Add support for x86_64 boot-stack VMs.
Adds a new --arch option to configure-seed-vm. This option is
passed into the seed.xml libvirt template to control the arch
being used.

Also, Updates the scripts/boot-elements script so that it
passes the correct architecture when creating the seed VM.

With this change I'm able to create x86_64 bootstrap VMs...