Fix call to install-dependencies.
If the current user running install-dependencies is not in the
libvirtd group, the user is added and install-dependencies is
reran.  This patch calls install-dependencies with $0 instead
of $PWD/$0.  There is no gaurantee that that the current working
directory is the scripts directory, so using $PWD/$0 could result
in an error.

Change-Id: I67e7eeaba365ded8fcaa1ba6b48d6759e90d5222