Add caching ability to boot-seed-vm
Currently boot-seed-vm creates a new image each time it is run, users
who want to restart a new identical seed VM shouldn't need to create it
again from scratch a second time.

This commit gives users the option to cache the seed image so it can be
reused again without a costly call to di-b to create a new one.

Change-Id: Ib8174227cf6e16ae59a64ef396b6a3c5f125be4e