Avoid using tempfs by default when not enough RAM avaiable
Building the seed image needs about 4.5Gb of space. disk-image-create will use a
tempfs on those system with more than 4GB of RAM, with an free space by default up to 50%
of the RAM available so those system with 8GB or less will fail to build the image.

Test that there is more than 9GB RAM avaible or avoid the use of tempfs to create the seed image.
This can be override by using the env variable DIB_NO_TMPFS=[0|1].

Change-Id: Iede5bf9d50f548805a79e72b3f8cf074bc3e6b3e