Convert devtest.md to devtest.sh.
We have an ongoing challenge with CI/toci : our user story
'devtest.md' is duplicated in the CI infrastructure. By using the well
known pattern of testable-documentation / documented tests we can have
cake and eat it too: a clear story that helps users understand whats
going on, and a script that can test our workflow end to end.

This will be a multi-part story to get the devtest aspects of toci
ported back into the main repository without compromising the docs -
this step sets up the framework, but devtest.sh isn't yet runnable.

Change-Id: If9c7a7b42c22a12799fed557117ad058f75ef6c2