Create endpoints on the seed with unset password.
As part of commit d44b07ea22ecb195b8aa950ebc7b3186442d5d9a,
setup-endpoints was updated to call register-endpoint with empty
passwords if they weren't specified on the command line.  The empty
password causes register-endpoint to generate a random password for
the endpoint.  However, devtest was not updated to create the
passwords before hand for the seed node like is shown for the
undercloud and overcloud.

I'm not sure if the intent was to continue to have the seed node use
unset passwords or update devtest to set passwords.  This change
updates devtest to use unset passwords, if that's not desired, I can
update it the other way instead.

Note this patch also requires https://review.openstack.org/#/c/46621/
which updates setup-endpoints to accept the *-password arguments.

Change-Id: I360ad9e7dae03e5ef1e36d3bb0eedcff12221bff