Shellify all of devtest.sh.
It is possible now to run a full devtest setup from the script. There
are still two points where the script is unable to detect a working
cloud node and so simply sleeps for 90 seconds. These should be urgently
replaced with appropriate duck tests.
We have also now inverted the greed in the image name globbing when
loading images into glance. This is necessary to support building in a
directory with a period in the name (i.e. ~/.cache/).

Change-Id: I75acd6cbc583b7c18a615ad72ee595977441df59