Script to save devtest environment
Add a write-tripleorc script that writes out the enviornment variables
that are set from the devtest instructions to a tripleorc file.  The
tripleorc file can then be sourced from a new shell if you happen to
have lost the shell you were running the instructions from.  Very
handy if say your suspended laptop eventually runs out of battery on a
long flight home :-).

Change-Id: I28aae9a1fae62019b8b7f8074586adc5f48ef366