Increase MTU on brbm (fixes iscsi packet loss)
iSCSI fails miserably when using an OVS bridge with the MTU
set to 1500. With it set to 1500 I repeatedly get iSCSI
connection errors in my iscsid log files.

This patch updates our setup-network script so that when we create brbm
we automatically set the MTU to 1504.

It also makes a note in devtest.md so that people using real bare metal
will know to increase their MTU's for the control plane to avoid
errors like this.

Change-Id: Iefdb0ab31769bc0a9cd15eab3948f3d78c35efdd