Register meaningful baremetal host names.
When registering baremetal hosts we used to hard code 'ubuntu'
everywhere.

With these two image element and heat template changes we can now set
the host names in nova.conf to 'seed' and 'undercloud' respectively:

 https://review.openstack.org/#/c/48497/
 https://review.openstack.org/#/c/48498/

This change updates register-nodes so that the $NODES variable is
used to create nova baremetal nodes. We already use this variable
to configure both the seed VM and undercloud so we may as well
make use of it for baremetal node names as well.

Change-Id: I9eab3b310c3eb7455438d58de26f67bd9db04e58