Add power management support to register nodes.
This changes makes it possible to pass real pm_address, pm_user, and
pm_password parameters to the nova baremetal-node-create commands.

To use simply define PM_IPS, PM_USERS, and PM_PASSWORDS variables
(space delimited) so that the associated user:password combination
correspond to the MACS (which we already require to use this script)

The motivation for this change is to make it possible to use this
script (and follow the dev/test story) using *real* hardware and
not just fake VMs.

Change-Id: I347edeb3b81473305d64f74ea2ac78d731361c9d