Register meaningful service node names.
When registering baremetal hosts we used to hard code 'ubuntu'
everywhere.

With these two image element and heat template changes we can now set
the host names in nova.conf to 'seed' and 'undercloud' respectively:

 https://review.openstack.org/#/c/48497/
 https://review.openstack.org/#/c/48498/

This change updates register-nodes (and setup-baremetal) so that
we pass in a $SERVICE_NODE parameter which is then used to register
baremetal nodes w/ Nova.

Change-Id: I9eab3b310c3eb7455438d58de26f67bd9db04e58