Swap service host names in setup-baremetal calls
Now we are using meaningful host names for seed and undercloud
machines, which is important for nova-compute to work properly.

setup-baremetal calls in devtest.sh seem to have swapped service
host names (when baremetal nodes are registered for undercloud
machine the 'seed' service host name is used and vice versa).

Change-Id: I1373eec38f28d6294f8d941f9b99f0781082a325