Redirect output of pool create call
When I ran create-nodes as instructed in devtest.md, I got back
a bogus MAC such that "SEED_MACS=Pool default defined".  It looks
like there was some output from the pool create call to virsh, so
redirect that command's output to /dev/stderr as is done for the
configure-vm call.

Change-Id: Ic3163178765a418b8c91a59ebbb04e18d236bcb7