Make pull-tools avoid offering to merge
In OpenStack we don't allow merge commits, and it probably isn't what
we would want even if we did allow them. Also, if we are on a branch,
we are going to assume the user does not want to even attempt a pull.

If the user is on master, and has diverged, we want to offer them a
likely next step, which is to rebase on origin/master.

Change-Id: I446394505d82bc30af6ead0d7dd1a6e11fe1e7b7