Setup overcloud passwords independently of undercloud passwords
Don't assume overcloud and undercloud can access the same file to get
the passwords from.  Add a script to create just the overcloud
passwords and use in devtest_overcloud.sh

Note: setup-passwords has not been changed, because it is outside of
what I am testing at this moment, but it should be adjusted to only
create the undercloud passwords.

Change-Id: I47de68d081ab3d8e83d014d416e229c055b42ec8