Switch from ">/dev/stderr" to ">&2"
Although the two should be equivalent (have the same effects) the way
they work is different. The ">/dev/stderr" does an open(), it goes
to the filesystem look for that file, opens it and write to it, it's
also subject to permission checks. The second way of doing it ">&2"
just uses file descriptor 2 so it doesn't hit the filesystem at all,
which would be safer and can't fail at permissions checks.

Change-Id: If9b0cb8dbeee4d29ba1387ba6f62a27223b409a1
Closes-Bug: #1236756