Setup CD overcloud public routing.
The CD overcloud requires non-devtest-default networking setup. The
tripleo-cd element needs to be able to control this: this patch seems
like a reasonable balance between JFDI and maintainable, at least for
now. In particular, the cd-tripleo element will own the config, so
it's not hidden outside of VCS and thus only-accessible-by-root.

Change-Id: Iea774b1b04d9d2d0fc159d39ef50592012caedef