Support openSUSE in install-dependencies
This should be enough to allow install-dependencies to run successfully
on recent versions of openSUSE, and has been tested on openSUSE 12.3.
It will not work for SLES due to the way the path to the
Cloud:OpenStack:Master repo is constructed, but it should be enough to
go on with for now.

Change-Id: I5b5485e644aa55ee96c835560a6b01869968136a