Enable dhcp-all-interfaces.
While we don't need this for VM's, it's needed for real hardware
where we cannot predict the network configuration of the machines
unless we want statically configured network rules in the images
or crufty manual configuration.

This is compatible with file injection - it just tries to ensure
all undefined interfaces have a configuration stanza.

We don't have an implementation yet for RPM based machines, but
they have a workaround via their netboot configuration so at
least for machines with an eth0 will be no worse off with this
enabled - and disabling file injection is a key step for scalability.

Change-Id: Ic3c48cd9c151276b7bd9771e21065dc48b114978