Setup admin and end user accounts in the overcloud.
While ongoing management is a matter of API use, we have a specific
use case for automation: we want controlled access to just TripleO
ATC's (vs self-signup via a portal), as part of the TripleO CD cloud.
I decided not to do something fancy for passwords (like GPG encrypting
to the user) and instead just capture an initial password for the user
on disk, which any admin can log in and give to the user, because
thats sufficient to get folk going. Until we have state preservation
we need to recreate the users each time, which is another reason to
keep the passwords around: that way it's not new every time for users.

Change-Id: Ie5b0f92f204af467fe3159cc2a05860d01608310