Build end user image during overcloud deploy.
This reduces overall cycle time by exploiting some natural
parallelism. On small machines it may reduce performance if we exceed
the concurrent available capacity, but most production environments
will have enough ram to be building in tmpfs, and 2 cores - enough for
one to build the image and one to be driving deployment logic.

Change-Id: I314c5784bba6cd4efdf7b4a80759db844857d0b0