Ensure the network is scheduled to an dhcp-agent
For some reason DHCP-agent is not present in neutron
agent-list output immediately after start, so we have
to wait a bit.

Neutron doesn't schedule new networks immediately, but
rather waits for the first port to be allocated on the
network. So we have a kind of a race here, because we
also create a neutron router which effectively triggers
port allocation and scheduling of the network. We should
be able to proceed in both cases: when the router was
created before we scheduled the network manually and vice
versa.

Partial-Bug: #1241042
Related-Bug: #1231366

Change-Id: Ib40183a21e6ca99fa46e5b94782535cad980eec2