Ensure the network is scheduled to an dhcp-agent
For some reason DHCP-agent is not present in neutron
agent-list output immediately after start, so we have
to wait a bit.

By the time we are going to schedule the neutron network
to the dhcp agent it might have already been scheduled,
so we should be able to handle this case too.

Partial-Bug: #1241042
Related-Bug: #1231366

Change-Id: Ib40183a21e6ca99fa46e5b94782535cad980eec2