Allow Neutron to schedule networks lazily
When a new network is created, Neutron won't schedule it
to one of the active DHCP agents until the first port on
this network is allocated.

This means that we should not bother to schedule the created
network manually and can just rely on Neutron for doing this
(actually, we've already been relying on Neutron for Seed VM,
since we switched to using of neutron-dhcp-agent).

Related-Bug: #1231366

Change-Id: I413407effe93312f2b5dd69d673590043becca33