Fix assigning of a floating IP to a Neutron port
In devtest.sh when a user VM is booted, we are trying to
assign a floating IP to a Neutron port before the latter
has been actually created. Creation succeeds, but the
floating IP is not bound to the port, so it's not usable.

After a request for booting a Nova instance has been issued,
we should wait before the corresponding Neutron port becomes
available (i.e. check neutron port-list output).

Closes-Bug: #1245480

Change-Id: I8ea9197471489cb536e00ed6d0f8d4cbb39b5e00