Add script to configure swift ring files
Adding a script "scripts/swift-init-ring" can be used to create a
ring file on multiple hosts. This script is not ideal as it creates
the ring file on each host. We can't create it on the host running
the script as we don't have access to swift-ring-builder, alternatively
I could install swift on this host but this didn't seem like a nice
alternative.

Change-Id: I21f77124b4edb34849e9b75ddea343c222709fa4