Fix floating ips when devtest.sh is hand executed
When devtest.sh is hand executed, the public interface of the
network node in overcloud is not added to br-ex bridge, which
makes floating ips unusable. This is caused by the fact that
we don't set the name of the public interface in devtest guide
(it's in devtest.sh, but it's not put to generated html doc).

Closes-Bug: #1252304

Change-Id: I9035b5ab26498954b46051f1b9bb982f0bc07c86