Use separate stack-create overcloud call in docs
Currently, devtest_overcloud doc contains stack-create overcloud
call, which accepts parameters, that we don't expose to users in
devtest story (e.g. NeutronPublicInterfaceDefaultRoute). It would
be less confusing to users, if we passed only parameters, that
are defined in devtest doc and rely on defaults of the overcloud
template for the rest of parameters.

Change-Id: Iec097bee09ff06997fcfd7a4ec4a3d639ad6c77f