Wait for nova to initialise
As the service dbsync's now happen in postconfigure.d nova is now
available some time after keystone. This commit adds a wait for the nova
api service in order to ensure we don't prematurely try to use nova.

This will be required once https://review.openstack.org/#/c/57466/ is
merged.

Change-Id: I6db65178b12df02b59c267b6d502c6bed1c6ac85