Add create-nodes option for baremetal interface
Add a final, optional argument to the create-nodes to define a specific
baremetal interface to be used and included as part of the name in libvirt.
This will allow us to define multiple TripleO test environments on different
bridges all on one machine/instance of libvirt.

So we would run:

create-nodes $NODE_CPU $NODE_MEM $NODE_DISK $NODE_ARCH 1 brbm1

And result would be:

baremetalbrbm1_0 and baremetalbrbm1-0.qcow2

Behavior without this argument remains the same.

Also update cleanup-env to include cleanup of alternate seed and baremetal
named VMs.

Change-Id: I111d51ef3a817655284a28faca2ccd211d2241cc