Make ext-net shared so users can use floatingips
Previously we kept ext-net admin-controlled, requiring admin privileges
to allocate floating ips. But the script does so at a time when we are
in a user account. Further, users will likely want to be able to
allocate their own floating ips.

We also now have to be more explicit when booting the demo instance,
because there are two networks available.

Change-Id: I7d07ae40689f5518b9d7df89b9ae1572d2c80c6d