Save the output from boot-seed-vm
We currently write the output of building the undercloud and overcloud
images to log files, this commit adds logging to the building of the
seed image.

The output from the disk-image-create contained in boot-seed-vm is
already saved to ~/.cache/image-create/image-file-cache but it doesn't
contain the output from all of boot-seed-vm and can be hard to find if
you don't know about it.

Change-Id: I3e75605fc889638fa77b1c1a5d46c6294fd101ea