Recommend using the stack user to ssh to the seed
We currently recommend using root to ssh to the seed, but using the
stack user works just fine too since the stackuser element will copy
your SSH key if the local-config element is enabled.

I think it's better to recommend people use the stack user because at
least that same user will work on the undercloud and overcloud, even
if it requires a password in those cases.

Basically, I'm finding myself randomly trying root, stack and heat-admin
and struggling to remember which user is appropriate for each node.

Change-Id: Iddeb47ebce6d0d14f2d2327724e288afe5f1e71d