Set correct permissions for ~/.ssh/authorized_keys
It's possible we end up creating ~/.ssh/authorized_keys in
install_dependencies, so we need to make sure it's permissions are set
to 0600.  Even if we didn't create the file, it's a change that can be
run without any pre-checks, as the previous "if" statement in the script
will create the file if it doesn't already exist.

Change-Id: I70038954a06deb56c4d1d602cb436b5bb563ae40