Fix wait for Neutron L2 Agent
A recent commit adds a wait for the Neutron L2 agent to be ready before
continuing with devtest. wait_for seem to have trouble with the way the
quotes are escaped. This commit changes them to wildcards instead.

The check was also waiting for the agent on the notcompute node,
according to the commit message the compute node was what was
intended(which makes more sense), this commit also changes the
check to wait for the agent on compute.

Change-Id: I231cde6263dd193767989f65b87162cf96a2803c