Add POWER_MANAGER support to devtest.
Adds an optional POWER_MANAGER option to devtest which can be
used to control the power manager used within the seed and undercloud
images. This can be useful when using the devtest scripts w/ real hardware.

This support already exists in tripleo-ci and once included in the devtest
scripts will make switching to the devtest scripts in tripleo-ci easier.

Change-Id: I976aa2f89e965dc9024e7967d7e593fb1f874000