Add devtest real BM MAC and PM command support
Add top level support for using real hardware with devtest.
This patch adds the ability to set MACS, PM_IPS, PM_USERS,
and PM_PASSWORDS to devtest.sh which is then used when
calling setup-baremetal.

If these variables are set there is no need to call create-nodes
for virtualized undercloud/overcloud machines so when MACS
is set those are selectively disabled.

The motivation with this commit is to reach parity with current
tripleo-ci features and eventually restructure tripleo-ci to
make use of the devtest scripts.

Change-Id: I41847e89c6b0c250c759f33182113af92e8a3c98