Permit setting the overcloud image update policy.
This is needed to preserve /mnt when rebuilding nodes. In principle we
can do stateful updates with rolling migrations etc - this is just
much quicker and is the path we have agreed on for getting out of the
quagmire we're in where noone can use use for production image based
updates.

Change-Id: I0efa8f52864f49ccdb885f6f655c732c951b3f7a