Set DIB_COMMON_ELEMENTS in devtest_overcloud.sh
For the cd-overcloud devtest_overcloud.sh is being run standalone
outside of devtest.sh. So this needs to be set seperately. With this
comes some duplication as the names of the common elements need to be
reset. This commit fixes the current error being seen but should be
followed up with something to remove the duplication.

Change-Id: Icf20b49956e43adb3631c2df491b3d98d9543c6b