Set DIB_COMMON_ELEMENTS in devtest_overcloud.sh
For the cd-overcloud devtest_overcloud.sh is being run standalone
outside of devtest.sh. So this needs to be set seperately. This should
be followed with a patch to t-i-e/elements/tripleo-cd/tripleo-cd.sh to
set an appropriate value.

Also revert the inclusion of the pypi-openstack and pip-cache elements
in images, so they can be discussed in another commit.

Change-Id: Icf20b49956e43adb3631c2df491b3d98d9543c6b