Start to split devtest_setup out more.
The goal of this effort is to be able to not create VM's during
devtest, so that CI can work with it's predefined environments.
Separately, for folk doing devtest steps by hand, having a safe thing
we can source, separate from the active script that installs packages
etc, seems like a good idea. Thus _testenv and _variables. Of course,
keeping the narrative intact is tricky too - so this patch just sets
up the layout, allowing me and StevenK to work in parallel on the
splitout.

Change-Id: I2cff0bfd09dcaff1e118f519b8d81a74496c6568