Use the JSON description to probe the seed VM.
This adds a dependency on jq to let us cleanly write autodetected seed
VM IP addresses into the test file, which will simplify things later
on in our workflow. Using the datafile is preferrable to complex
checks in devtest which is afterall meant to be as much manual as
script.

Change-Id: I81e49d16077c564dd1e5d3e6f6e7a26e7c56ef5c