Permit deploying an overcloud with SSL endpoints.
SSL is a key feature in todays environment. This patch matches what
Chris Jones did by hand for the currrent cd-undercloud - it may well
be something we evolve in future.

Some services are not yet running with SSL configuration, I've
commented around them appropriately and we can iterate on making them
work.

Change-Id: I511d5f5a3bad3169fc1b61dce5d3d055f1cd5960