Change communication with remote host
There was a mismatch between the commands that were called in
scripts/boot-seed-vm and the commands expected by the test environment.
o Each TE has a number on the Host which should be used to select the
environment to use.
o ci_commands isn't expecting us to use rsync, rather we should pipe the
seed qcow2 image into a custom command "copyseed".
o ci_commands is expecting the virsh commands to be run(not a qemu+ssh
connection). Running the virsh commands seems better as it allows us to be
more granular with our filter. qemu+ssh just ends up calling
> nc -U /var/run/libvirt/libvirt-sock'
which presumably would give the user untethered access to virsh.

Change-Id: Id99d05afa65bee7eb357f8b28fd70ed3aaa014f4