Fix SSL .crt error when not using SSL.
Fix an issue with 1845a169f9f38597e86d77ceee93be9f82f1713a where
SSLBASE was required to be set in order to use devtest_overcloud.sh.

This patch adds a check for SSLBASE variable and only tries to load the
associated .crt and .key files if it is set.

Change-Id: I09e800e78401ade52a0a21df7e33956c055e2c88
Closes-bug: 1271639