Devtest Overcloud all-in-one support.
If OVERCLOUD_ALL_IN_ONE is set builds the overcloud
by using a single image. This is useful if you want
to save memory or only have access to a single
bare metal overcloud machine.

This is something we've supported in the old TOCI scripts and
it has proven useful to some of us.

For now the all-in-one commands are all undocumented.

Change-Id: I392322b79b4635d77047f020bbd40e4a4676ac28