Stop sourcing devtest_seed.sh.
The only thing we use was the noproxy setting and keystone creds -
neither of which is really semantic semantic, so just duplicating it
outside the script is easy and IMO reasonably tasteful.

An alternative approach would be to source seedrc at the top of
devtest_undercloud but that would make using a different seed harder
and for devtest_overcloud we will want to avoid that because folk
doing overcloud only will likely want to skip the undercloud
altogether.

Change-Id: If05e5686425da3cb20c94cfff40b16da25cfc65c