Ensure libvirtd is always configured on Red Hat
The code that populates libvirtd.conf on Red Hat operating systems was
scoped under a conditional that checks for user's presence in the
libvirtd group. This meant that when the libvirtd group already
existed and user was a member of that group, libvirtd.conf would not
get populated.

Move out the code that populates libvirtd.conf to top level (it is
guarded by grep anyway), and also move out the code that ensures
libvirtd group exists before it is used in libvirtd configuration.

Minor cleanup: "grep ... > /dev/null" changed to "grep -q ..."

Change-Id: Ie81a2ac883304e546fc95e8b5a2201bcaea9fc9c