Messaging backend selection
There is another patch that removes rabbitmq-server from
boot-stack's element dependencies.

https://review.openstack.org/#/c/72516/

This patch should be deployed before 72516.

This patch selects rabbitmq-server as the default messaging
backend and sets it in a MESSAGING_BACKEND environment
variable. This patch provides a way for users to select
a different backend if desired.

qpidd is an alternate backend. See the README in the qpidd
image element for details. Right now you do need to edit
the seed's config.json and heat templates in order to use
qpidd.

Change-Id: Ic19af7f6071af04bd6bbce8196bf80ab2dae6840