Switch default nic driver to virtio
This was made configurable some time ago to give people time to ensure
using virtio doesn't cause problems.

Since then I've been using virtio exclusively (most of the time on
Fedora) and have verified multiple devtest combinations including
Fedora 20 devtest environment host
   Fedora 20 devtest instances
   Ubuntu Saucy devtest instances
   Ubuntu Saucy Seed booting a Ubuntu Precise bm instance
Ubuntu Saucy devtest environment host
   Fedora 20 devtest instances
   Ubuntu Saucy devtest instances
   Fedora 20 Seed booting a Ubuntu Precise bm instance

The overcloud ci job (back when it included the undercloud),and
presumably devtest ran 15 minutes faster with virtio nodes so its time
to make the switch.

I believe the reason this was set to e1000 in the first place is because
a old version of dhclient on older ubuntu cloud images dropped dhcp
packets without a checksum. We no longer target this image in tripleo
and according to the changelogs for the precise packaging the fix was
included in isc-dhcp-client 4.1.ESV-R4-0ubuntu5.8. The testing
described above verified its working as expected.

Of course a new set of test environments will need to be built once
this is merged in order for CI to take advantage.

Change-Id: I41046acca22164619bae83ec5cb1d22d1ddf2946