Switch default nic driver to virtio
This was made configurable some time ago to give people time to ensure
using virtio doesn't cause problems.

Since then I've been using virtio exclusively (most of the time on
Fedora) and have verified the ubuntu based ci work with virtio.

The overcloud ci job(and presumably devtest) runs 15 minutes faster
with virtio nodes so its time to make the switch.

I believe the reason this was set to e1000 in the first place is because
a old version of dhclient on older ubuntu cloud images dropped dhcp
packets without a checksum. We no target this image in tripleo.

Of course a new set of test environments will need to be built once
we're happy this doesn't cause any problems.

Change-Id: I41046acca22164619bae83ec5cb1d22d1ddf2946