Increase time waiting for nova to initialize
Various parts of the boot process take longer on the ci test
environments when running multiple tests on the same host
simultaneously, specifically the various db syncs seem to take upto 4
times as long(when comparing to a local test). To allow for this we
give nova some more time to be initialized.

Long term we should spend some time investigating what the bottleneck is
to improve the performance.

Partial-Bug: #1284054
Change-Id: I3df6f24fd906e0fd74c36e2ba96e2b731e60416b