Add mechanism for wait_for command to bail out on an unrecoverable error
devtest will wait for a long time (3000 seconds for an overcloud) even
if the stack has failed and will never be ready.

This patch updates wait_for so that if the COMMAND fails with a status
greater then 1 then we exit the wait_for command with a error status of 1
notifying the user.

Updated stack-ready to use this. If the stack status becomes CREATE_FAILED
then we exit with status 2 exiting the long wait_for loop.

Change-Id: Iac0d49754e2a7059784ca29773056e13b42ad88a