Add FAIL_MATCH_OUTPUT to wait_for...
Adds a new ENV option that can be used with the wait_for command
to fail fast if a supplied regex string is found in the commands
stdout result.

As part of this change we are also redirectly stderr for the command
into the OUTPUT variable so it can also be checked.

Change-Id: I3e6b76bd6da66da3afb92059b757da18c9b06497