Cause wait_for heat stack-create fail fast
Add a script called wait_for_stack_ready that calls heat stack-show in a
wait_for loop waiting for either the stack to be ready or else when the stack
fails and will never be ready.

wait_for_stack_ready sets up the FAIL_MATCH_OUTPUT environment variable to
check when stack-create fails and setups a new SUCCESSFUL_MATCH_OUTPUT that
check the output from the heat stack-show for when the stack reaches the
UPDATE_COMPLETE or CREATE_COMPLETE state.

Change-Id: I7daf5010d15e458d98698fb8b7e744e2a17596b1