Add standalone script to build all images.
Adding a new script to generate all images: seed, undercloud,
overcloud-control/compute and user image:

This has been used to generate a set of images and repeatedly
verified that a devtest.sh -c run with this set of images
successfully brings up tripleo stack (demo vm is sshable).

- It is useful to be able to generate images offline for consumption
by other users in bringing up a Tripleo stack (e.g. for testing),
without going through the steps for bringing up the stack.
- This is also intended to make the composition of elements for
each Tripleo element clearer and defined in one place.
- Initially pushing this as a standalone script for feedback. The
ultimate aim is to refactor devtest.sh to call this script and remove
the image-building sections from devtest_seed/undercloud/overcloud.sh.
- Currently, devtest_images takes a cache parameter to skip image
building and will also always generate an image if not present.
This was done to match current devtest behaviour. When
the above refactoring is done, it will make more sense for cache check
to be made in devtest.sh, where it can skip calling devtest_images --
in that case, another parameter may be required to only rebuild if not
present.

Change-Id: Iddf06c79cd7b33cc671e5775da32b2037fa5e580