Use the DIB_REPOLOCATION / DIB_REPOREF for the bootstrap.
The bootstrap process hard-codes the repos it checks out for
tripleo-incubator, tripleo-image-elements, diskimage-builder,
and tripleo-heat-templates.

This adds support for DIB_REPOLOCATION / DIB_REPOREF variables
to the bootstrap process.

Although the environment variables might be technically misnamed
(since they're used by the bootstrap process, before DIB gets a
look-in), I picked those so that they'll work with any environment-
setting script that trawls a directory of checkouts to set these up
prior to invoking devtest.sh.

This is just a small convenience whilst working on the incubator
itself.

Change-Id: I33e16cb3946519a686c1692f5e2d39424c13bcbf