Wait for baremetal nodes before creating undercloud stack.
devtest_undercloud can fail due to attempting to bring up heat
stack before baremetal nodes are available in the seed. This
issue is typically triggered when devtest is run with cached
images, as the undercloud stack creation occurs soon after
baremetal node setup in the seed in that scenario. This change
introduces a wait for baremetal node to be available before
bringing up undercloud heat stack.

Closes-Bug: #1288199
Change-Id: I720ebf1333607b0daa2f8989f7e31d02f7903ef3