Give the user instructions they can follow instead of $(dirname $0)
We have two remaining instances of $(dirname $0) that get shown to the
user in the instructions for ru nning the devtest sub-scripts by hand.

One of these is obsolete; by that point, $TRIPLEO_ROOT is set and in
$PATH, so we can run the script directly. I've removed the $(dirname $0)
as it isn't needed.

The one case we do have to care about the path is the initial source
devtest_variables. I can't see a way to have a single command-line that
works for both devtest.sh and the user running sub-scripts manually, so
I've manually duplicated that line in a slightly different format so
that the docs will show the user a useful line.

Change-Id: Ib5fbaff5edd013bcba9c672f129ed14c54ff2048