Remove $(dirname $0) from devtest.sh docs.
We have two remaining instances of $(dirname $0) that get shown to the
user in the instructions for running the devtest sub-scripts by hand.

When these commands are run inside devtest.sh, they locate the directory
that devtest.sh is in, so they serve as an fixed reference point from
which we can locate the other scripts that need to be run.

When a user runs these by hand, they don't get anything useful. The
intent here is to replace the unhelpful instructions with a useful
equivalent - something the user can actually enter that will have the
desired effect.

One of these is obsolete; by that point, $TRIPLEO_ROOT is set and in
$PATH, so we can run the script directly. I've removed the $(dirname $0)
as it isn't needed.

The one case we do have to care about the path is the initial source
devtest_variables. I can't see a way to have a single command-line that
works for both devtest.sh and the user running sub-scripts manually, so
I've manually duplicated that line in a slightly different format so
that the docs will show the user a useful line.

Change-Id: Ib5fbaff5edd013bcba9c672f129ed14c54ff2048