Revert "Remove $(dirname $0) from devtest.sh docs."
This reverts commit 3b2bdd757d6307741ea1e518d823714838dcca23.

On a brand new machine, this breaks as devtest_setup.sh is not in the
path - $TRIPLEO_ROOT/tripleo-incubator/scripts does not exist until
after we run devtest_setup.sh, so we cannot run devtest_setup.sh from
that directory.

This works on any machine that already has a checkout in
$TRIPLEO_ROOT/triple-incubator.

This slipped through CI because, as part of the gate setup, all the
relevant repositories are checked out - so the first time
devtest_setup.sh runs in CI, $TRIPLEO_ROOT/tripleo-incubator/scripts
exists. This is not true for a user who is following our docs.

Change-Id: I4f842a91455fa6a08a3004ff177b3e94452872ae