Set default values for repo location and ref
Steal code from diskimage-builder's source-repositories element to make
sure that we never reference an unbound variable.

First we populate the name of a potential variable into an _OVERRIDE
variable; then we set the real variable to either the contents of a
variable with that name or the default.

This avoids any chance of ever referencing an unbound variable, even in
an eval.

Also, change the message displayed in the case that the tool is already
checked out, but not on the master branch, and DIB_REPOREF_tool is not
set, to be clearer about why we're skipping the pull of that repo

Change-Id: Ib98b20dbacfc8cdf948bee1265adc9485894c73f
Closes-Bug: #1294477