Set default values for repo location and ref
Steal code from diskimage-builder's source-repositories element to make
sure that we never reference an unbound variable.

First we populate the name of a potential variable into an _OVERRIDE
variable; then we set the real variable to either the contents of a
variable with that name or the default.

This avoids any chance of ever referencing an unbound variable, even in
an eval.

Change-Id: Ib98b20dbacfc8cdf948bee1265adc9485894c73f
Closes-Bug: #1294477