Make the baremetal-network configurable
Adds support for a new (optional) "baremetal-network" config
section to the testenv.json file. This JSON metadata can be used
to specify a custom baremetal network other than the hard coded
192.0.2.0/24 we have been using up until this point.

This effectively makes the flat network ranges used to bootstrap
the seed and undercloud fully configurable and can be very
useful when using the devtest scripts on real baremetal
(like some of us do for our CI/CD environments).

Change-Id: I6e89dc985fdd276a05f0e82c3905cc5efe646a4c